requirements.txt
aiogram~=3.25.0
fastapi~=0.133.1
uvicorn~=0.41.0
python-dotenv~=1.2.1